Enemies now take multiple hits to kill, and a health bar appears after they get damaged to show the player how much damage they're doing.

+++ Enemy.pde
@@ -1,4 +1,9 @@
 class Enemy extends Actor {
+
+  static final int MAX_HP = 100;
+  final color HP_BAR_COLOR = color(230, 20, 20);
+  
+  int health;
   
   public Enemy() {
     this(0, 0);
@@ -7,6 +12,7 @@ class Enemy extends Actor {
   public Enemy(float x, float y) {
     super(x, y);
     this.fillColor = color(10, 10, 10);
+    this.health = MAX_HP;
   }
   
   void setVelocity() {
@@ -21,7 +27,26 @@ class Enemy extends Actor {
   
   void handleEntityCollision(Entity other) {
     if (other instanceof Bullet) {
-      this.delete();
+      health -= 20;
+      if (health <= 0) {
+        this.delete();
+      }
+    }
+  }
+  
+  void drawHealthBar() {
+    stroke(255);
+    fill(0);
+    rect(x - 8, y - 18, width + 16, 13);
+    noStroke();
+    fill(HP_BAR_COLOR);
+    rect(x - 6, y - 16, (width + 13) * (float) health / MAX_HP, 10);
+  }
+  
+  void draw() {
+    super.draw();
+    if (health < MAX_HP) {
+      drawHealthBar();
     }
   }
 }
